Refactor NotesList
